Rename ResourceStatement.g4 to CommonDistSQLStatement.g4
