+++ shardingsphere-distsql-parser/shardingsphere-distsql-parser-engine/src/main/antlr4/imports/Keyword.g4
@@ -35,14 +35,10 @@ DROP
     : D R O P
     ;
 
-PREVIEW
-    : P R E V I E W
-    ;
-     
 SET
     : S E T
     ;
-    
+
 SHOW
     : S H O W
     ;
