+++ shardingsphere-distsql-parser/shardingsphere-distsql-parser-engine/src/main/java/org/apache/shardingsphere/distsql/parser/api/DistSQLStatementParserEngine.java
@@ -18,8 +18,7 @@
 package org.apache.shardingsphere.distsql.parser.api;
 
 import org.antlr.v4.runtime.misc.ParseCancellationException;
-import org.apache.shardingsphere.distsql.parser.core.administration.SCTLStatementParserEngine;
-import org.apache.shardingsphere.distsql.parser.core.resource.ResourceSQLStatementParserEngine;
+import org.apache.shardingsphere.distsql.parser.core.common.CommonDistSQLStatementParserEngine;
 import org.apache.shardingsphere.distsql.parser.core.rule.RuleSQLStatementParserEngine;
 import org.apache.shardingsphere.sql.parser.exception.SQLParsingException;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.SQLStatement;
@@ -37,13 +36,9 @@ public final class DistSQLStatementParserEngine {
      */
     public SQLStatement parse(final String sql) {
         try {
-            return new ResourceSQLStatementParserEngine().parse(sql);
+            return new CommonDistSQLStatementParserEngine().parse(sql);
         } catch (final ParseCancellationException | SQLParsingException ignored) {
-            try {
-                return new RuleSQLStatementParserEngine().parse(sql);
-            } catch (final SQLParsingException ignoredToo) {
-                return new SCTLStatementParserEngine().parse(sql);
-            }
+            return new RuleSQLStatementParserEngine().parse(sql);
         }
     }
 }
