+++ shardingsphere-distsql-parser/shardingsphere-distsql-parser-engine/src/main/java/org/apache/shardingsphere/distsql/parser/core/common/CommonDistSQLParser.java
@@ -15,13 +15,25 @@
  * limitations under the License.
  */
 
-grammar ResourceStatement;
+package org.apache.shardingsphere.distsql.parser.core.common;
 
-import Symbol, RDLStatement, RQLStatement;
+import org.antlr.v4.runtime.TokenStream;
+import org.apache.shardingsphere.distsql.parser.autogen.CommonDistSQLStatementParser;
+import org.apache.shardingsphere.sql.parser.api.parser.SQLParser;
+import org.apache.shardingsphere.sql.parser.api.visitor.ASTNode;
+import org.apache.shardingsphere.sql.parser.core.ParseASTNode;
 
-execute
-    : (addResource
-    | dropResource
-    | showResources
-    ) SEMI?
-    ;
+/**
+ * SQL parser for common dist SQL.
+ */
+public final class CommonDistSQLParser extends CommonDistSQLStatementParser implements SQLParser {
+    
+    public CommonDistSQLParser(final TokenStream input) {
+        super(input);
+    }
+    
+    @Override
+    public ASTNode parse() {
+        return new ParseASTNode(execute());
+    }
+}
